added the logic for Neural Network
